+++ src/lib.rs
@@ -74,7 +74,6 @@ pub mod imex;
 mod scheduler;
 #[macro_use]
 mod job;
-mod format_flowed;
 pub mod key;
 mod keyring;
 pub mod location;